[LASB-3000] Move the external mapping to a new 'means.json' to match the schema on laa-criminal-legal-aid-schemas.
