+++ crime-applications-adaptor/src/main/java/uk/gov/justice/laa/crime/applications/adaptor/mapper/crimeapply/DependantsMapper.java
@@ -1,7 +1,7 @@
 package uk.gov.justice.laa.crime.applications.adaptor.mapper.crimeapply;
 
 import uk.gov.justice.laa.crime.applications.adaptor.model.crimeapplicationsadaptor.common.Dependants;
-import uk.gov.justice.laa.crime.applications.adaptor.model.criminalapplicationsdatastore.Dependant;
+import uk.gov.justice.laa.crime.applications.adaptor.model.criminalapplicationsdatastore.maat.Dependant;
 
 import java.util.ArrayList;
 import java.util.Iterator;
